Rename to fourierlayer.py
